pyproject.toml
[project]
name = "grid-trader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.11",
    "ccxt>=4.4.88",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "passlib>=1.7.4",
    "psutil>=7.0.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "tenacity>=9.1.2",
    "uvicorn>=0.34.3",
    "websockets>=15.0.1",
]
